Move MenuBuilder from Util/MenuBuilder to Component/MenuBuilder
